Cleanup render utility
